Purge old DELETEs during live compaction
This change is to address Issue 112, 'Old deletes need to be purged
during Compaction'.

It adds the ability to specify a
sirius.log.compaction-max-delete-age-hours configuration property to
cause old Delete events to be purged during live compaction.

The configuration property defaults to a very large value, so by default
Delete events will not be purged.

It also adds an option to WalTool to purge old Deletes, along with
other options for old PUTs.